requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
cryptography==35.0.0
h11==0.12.0
helium==3.0.8
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-failed-screenshot==1.0.2
pytest-rerunfailures==10.2
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
